fix all the nonsense
