+++ napari/components/experimental/chunk/__init__.py
@@ -1,3 +1,13 @@
 """chunk module"""
 from ._loader import chunk_loader, synchronous_loading, wait_for_async
-from ._request import ChunkLocation, ChunkRequest, LayerRef
+from ._request import ChunkLocation, ChunkRequest, LayerRef, OctreeLocation
+
+__all__ = [
+    'ChunkLocation',
+    'OctreeLocation',
+    'ChunkRequest',
+    'LayerRef',
+    'chunk_loader',
+    'wait_for_async',
+    'synchronous_loading',
+]
