+++ napari/components/experimental/chunk/_request.py
@@ -66,6 +66,52 @@ class ChunkLocation:
         return cls(LayerRef.from_layer(layer))
 
 
+class OctreeLocation(ChunkLocation):
+    """Location of one chunk within the octree.
+
+    Parameters
+    ----------
+    layer_ref : LayerRef
+        Referen to the layer this location is in.
+    slice_id : int
+        The id of the OctreeSlice we are in.
+    level_index : int
+        The octree level index.
+    row : int
+        The chunk row.
+    col : int
+        The chunk col.
+    """
+
+    def __init__(
+        self,
+        layer_ref: LayerRef,
+        slice_id: int,
+        level_index: int,
+        row: int,
+        col: int,
+    ):
+        super().__init__(layer_ref)
+        self.slice_id: int = slice_id
+        self.level_index: int = level_index
+        self.row: int = row
+        self.col: int = col
+
+    def __str__(self):
+        return f"location=({self.level_index}, {self.row}, {self.col}) "
+
+    def __eq__(self, other) -> bool:
+        return (
+            self.slice_id == other.slice_id
+            and self.level_index == other.level_index
+            and self.row == other.row
+            and self.col == other.col
+        )
+
+    def __hash__(self) -> int:
+        return hash((self.slice_id, self.level_index, self.row, self.col))
+
+
 class ChunkRequest:
     """A request asking the ChunkLoader to load data.
