+++ napari/layers/image/experimental/__init__.py
@@ -1,5 +1,5 @@
 """layers.image.experimental
 """
-from .octree_chunk import OctreeChunk, OctreeChunkGeom, OctreeLocation
+from .octree_chunk import OctreeChunk, OctreeChunkGeom
 from .octree_intersection import OctreeIntersection
 from .octree_level import OctreeLevel
