+++ napari/layers/image/experimental/_tests/test_octree_import.py
@@ -1,0 +1,7 @@
+import subprocess
+import sys
+
+
+def test_octree_import():
+    cmd = [sys.executable, '-c', 'import napari; v = napari.Viewer()']
+    subprocess.run(cmd, check=True, env={'NAPARI_OCTREE': '1'})
