+++ napari/layers/image/image.py
@@ -1,7 +1,10 @@
 """Image class.
 """
+from __future__ import annotations
+
 import types
 import warnings
+from typing import TYPE_CHECKING
 
 import numpy as np
 from scipy import ndimage as ndi
@@ -18,13 +21,8 @@ from ._image_slice import ImageSlice
 from ._image_slice_data import ImageSliceData
 from ._image_utils import guess_multiscale, guess_rgb
 
-# Use special ChunkedSlideData for async.
-if config.async_loading:
-    from .experimental._chunked_slice_data import ChunkedSliceData
-
-    SliceDataClass = ChunkedSliceData
-else:
-    SliceDataClass = ImageSliceData
+if TYPE_CHECKING:
+    from ...components.experimental.chunk import ChunkRequest
 
 
 # It is important to contain at least one abstractmethod to properly exclude this class
@@ -592,10 +590,21 @@ class _ImageBase(IntensityVisualizationMixin, Layer):
             thumbnail_source = None
 
         # Load our images, might be sync or async.
-        data = SliceDataClass(self, image_indices, image, thumbnail_source)
+        data = self._SliceDataClass(
+            self, image_indices, image, thumbnail_source
+        )
         self._load_slice(data)
 
-    def _load_slice(self, data: SliceDataClass):
+    @property
+    def _SliceDataClass(self):
+        # Use special ChunkedSlideData for async.
+        if config.async_loading:
+            from .experimental._chunked_slice_data import ChunkedSliceData
+
+            return ChunkedSliceData
+        return ImageSliceData
+
+    def _load_slice(self, data: ImageSliceData):
         """Load the image and maybe thumbnail source.
 
         Parameters
@@ -610,7 +619,7 @@ class _ImageBase(IntensityVisualizationMixin, Layer):
             # property is now false, since the load is in progress.
             self.events.loaded()
 
-    def _on_data_loaded(self, data: SliceDataClass, sync: bool) -> None:
+    def _on_data_loaded(self, data: ImageSliceData, sync: bool) -> None:
         """The given data a was loaded, use it now.
 
         This routine is called synchronously from _load_async() above, or
@@ -755,7 +764,6 @@ class _ImageBase(IntensityVisualizationMixin, Layer):
 
     # For async we add an on_chunk_loaded() method.
     if config.async_loading:
-        from ...components.experimental.chunk import ChunkRequest
 
         def on_chunk_loaded(self, request: ChunkRequest) -> None:
             """An asynchronous ChunkRequest was loaded.
@@ -766,7 +774,7 @@ class _ImageBase(IntensityVisualizationMixin, Layer):
                 This request was loaded.
             """
             # Convert the ChunkRequest to SliceData and use it.
-            data = SliceDataClass.from_request(self, request)
+            data = self._SliceDataClass.from_request(self, request)
             self._on_data_loaded(data, sync=False)
